+++ CHANGES.rst
@@ -29,6 +29,9 @@ Specviz2d
 - Improved logic for initial guess for position of "Manual" background trace in spectral extraction
   plugin. [#1738]
 
+- Now supports loading a specific extension of the 2D spectrum file and 
+  transposing data on load. [#1705]
+
 - Spectral extraction plugin now supports visualizing and exporting the 1D spectrum associated
   with the background region. [#1682]